pyproject.toml
[project]
name = "esoteric-vectors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.5",
    "langchain[google]>=0.3.25",
    "langchain-community>=0.3.0",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "langmem>=0.0.27",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "typing-extensions>=4.14.0",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "requests>=2.32.0",
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",
    "langchain-openai>=0.2.0",
    "pyephem>=4.1.5",
    # Auth0 dependencies
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.25.0",
    # Stripe integration
    "stripe>=8.10.0",
    # Testing dependencies
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
]
